+++ packages/cubejs-testing/test/smoke-redshift.test.ts
@@ -2,7 +2,13 @@ import cubejs, { CubejsApi } from '@cubejs-client/core';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { afterAll, beforeAll, jest } from '@jest/globals';
 import { BirdBox, getBirdbox } from '../src';
-import { DEFAULT_API_TOKEN, DEFAULT_CONFIG, testQueryMeasure } from './smoke-tests';
+import {
+  DEFAULT_API_TOKEN,
+  DEFAULT_CONFIG,
+  JEST_AFTER_ALL_DEFAULT_TIMEOUT,
+  JEST_BEFORE_ALL_DEFAULT_TIMEOUT,
+  testQueryMeasure,
+} from './smoke-tests';
 
 // Suite that can be re-used with different ENV configs
 const reusableSuite = (env: { [key: string]: string} = {}) => () => {
@@ -29,14 +35,14 @@ const reusableSuite = (env: { [key: string]: string} = {}) => () => {
     client = cubejs(async () => DEFAULT_API_TOKEN, {
       apiUrl: birdbox.configuration.apiUrl,
     });
-  });
+  }, JEST_BEFORE_ALL_DEFAULT_TIMEOUT);
 
   afterAll(async () => {
     await birdbox.stop();
     Object.keys(env).forEach((key) => {
       delete process.env[key];
     });
-  });
+  }, JEST_AFTER_ALL_DEFAULT_TIMEOUT);
 
   test('query measure', () => testQueryMeasure(client));
 };